requirements.txt
asgiref==3.7.2
Django==5.0.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1


django-filter~=23.5